+++ ArgumentResolver/ArgumentResolver.php
@@ -145,10 +145,12 @@ final class ArgumentResolver implements ArgumentResolverInterface
         $builtinTypeResolver = new Resolver\BuiltinTypeValueResolver();
         $backedEnumResolver = new Resolver\BackedEnumValueResolver();
         $dateTimeResolver = new Resolver\DateTimeValueResolver();
+        $inputFileResolver = new Resolver\InputFileValueResolver();
 
         return [
             $backedEnumResolver,
             new Resolver\UidValueResolver(),
+            $inputFileResolver,
             $builtinTypeResolver,
             new Resolver\MapInputValueResolver($builtinTypeResolver, $backedEnumResolver, $dateTimeResolver),
             $dateTimeResolver,